rpclib index

Library rpclib

The entry point of this library is the module: Rpclib.

Library rpclib.cmdliner

The entry point of this library is the module: Cmdlinergen.

Library rpclib.core

This library exposes the following toplevel modules:

Codegen

Idl The Idl module is for declaring the types and documentation for RPC calls

Pythongen

Rpc

Rpc_genfake

Rpcmarshal

Library rpclib.internals

The entry point of this library is the module: Internals.

Library rpclib.json

The entry point of this library is the module: Jsonrpc.

Library rpclib.markdown

The entry point of this library is the module: Markdowngen.

Library rpclib.xml

The entry point of this library is the module: Xmlrpc.